Add segment CDP logic to send_events.py
