+++ pyproject.toml
@@ -42,6 +42,8 @@ python-dotenv = "^0.21.0"
 ruff = "^0.1.3"
 pandas = "^2.1.2"
 faker = "^19.13.0"
+amplitude-analytics = "^1.1.4"
+rudder-sdk-python = "^2.0.2"
 
 [tool.pytest.ini_options]
 minversion = "6.0"
